requirements.txt
allure-behave==2.9.41
allure-pytest==2.8.40
allure-python-commons==2.9.41
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.9.3
behave==1.2.6
browserstack-local==1.2.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
chromedriver-autoinstaller==0.2.2
colorama==0.4.4
et-xmlfile==1.0.1
geckodriver-autoinstaller==0.1.0
idna==2.10
iniconfig==1.1.1
jsonpath==0.82
lxml==4.6.3
openpyxl==3.0.7
packaging==20.9
parse==1.19.0
parse-type==0.5.2
Paver==1.3.4
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==6.2.4
requests==2.25.1
robotframework==4.0.3
robotframework-pabot==2.0.0
robotframework-pythonlibcore==2.2.1
robotframework-seleniumlibrary==5.1.3
robotframework-stacktrace==0.4.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tqdm==4.60.0
urllib3==1.26.5
webdrivermanager==0.10.0
xlrd==2.0.1
xlwt==1.3.0
